Modified the code to address conversion issues
Since the Tweetinvi way to do things is complicated, i decided to create
a different way to deal with the user time line.
This way the control in the gathering process is all mine.

However, i wanted to convert the tweets received to tweetinvi tweets (so
i don't have to deal with 2 kinds of tweets), so i created a converter
inside the tweetinvi code (gets most properties except the entities -
hashtags and such).

Also, added some non-functional methods (since at first i thought of
doing it the right way, using Tweetinvi code and style)